+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## 0.4.13 - 
+### Added
+- utils添加conform-dlg函数
 
 ## 0.4.12 - 2020-11-30
 ### Added